Use double instead of float

+++ 03_sum_arrays.cu
@@ -3,25 +3,25 @@
 #include <string.h>
 #include <time.h>
 
-void sum_arrays_on_host( const int N, float *A, float *B, float *C )
+void sum_arrays_on_host( const int N, double *A, double *B, double *C )
 {
     for( int idx = 0; idx < N; idx++ ) {
         C[idx] = A[idx] + B[idx];
     }
 }
 
-void initial_data( int Ndata, float *dat )
+void initial_data( int Ndata, double *dat )
 {
     // Generate different seed for random number
     time_t t;
     srand( (unsigned int) time(&t) );
 
     for( int i = 0; i < Ndata; i++ ) {
-        dat[i] = (float)( rand() & 0xFF )/10.0f;
+        dat[i] = (double)( rand() & 0xFF )/10.0f;
     }
 }
 
-void print_data( int Ndata, float *A )
+void print_data( int Ndata, double *A )
 {
     for( int i = 0; i < Ndata; i++ ) {
         printf("%8d %18.10f\n", i, A[i]);
@@ -31,13 +31,13 @@ void print_data( int Ndata, float *A )
 int main( int argc, char **argv)
 {
     int Ndata = 10;
-    size_t Nbytes = Ndata * sizeof(float);
+    size_t Nbytes = Ndata * sizeof(double);
 
-    float *h_A, *h_B, *h_C;
+    double *h_A, *h_B, *h_C;
 
-    h_A = (float*) malloc(Nbytes);
-    h_B = (float*) malloc(Nbytes);
-    h_C = (float*) malloc(Nbytes);
+    h_A = (double*) malloc(Nbytes);
+    h_B = (double*) malloc(Nbytes);
+    h_C = (double*) malloc(Nbytes);
 
     initial_data( Ndata, h_A );
     initial_data( Ndata, h_B );
